+++ cmd/ots/client_test.go
@@ -7,18 +7,6 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-type FakeClientConfig struct{}
-
-func (f FakeClientConfig) NewClient() (Client, error) { return &FakeClient{}, nil }
-
-type FakeClient struct {
-	Client
-}
-
-func (f FakeClient) Organizations() tfe.Organizations { return &FakeOrganizationsClient{} }
-
-func (f FakeClient) Workspaces() tfe.Workspaces { return &FakeWorkspacesClient{} }
-
 func TestClientSanitizeAddress(t *testing.T) {
 	tests := []struct {
 		name    string
