+++ cmd/ots/workspace.go
@@ -1,10 +1,8 @@
 package main
 
 import (
-	"context"
 	"os"
 
-	"github.com/hashicorp/go-tfe"
 	"github.com/spf13/cobra"
 )
 
@@ -19,29 +17,7 @@ func WorkspaceCommand() *cobra.Command {
 	cmd.Flags().StringVar(&cfg.Token, "token", os.Getenv("OTS_TOKEN"), "Authentication token")
 
 	cmd.AddCommand(WorkspaceLockCommand(&cfg))
-	//cmd.AddCommand(WorkspaceUnlockCommand(&cfg))
+	cmd.AddCommand(WorkspaceUnlockCommand(&cfg))
 
 	return cmd
-}
-
-type FakeWorkspacesClient struct {
-	tfe.Workspaces
-}
-
-func (f *FakeWorkspacesClient) Read(ctx context.Context, org string, ws string) (*tfe.Workspace, error) {
-	return &tfe.Workspace{
-		ID: "ws-123",
-	}, nil
-}
-
-func (f *FakeWorkspacesClient) Lock(ctx context.Context, id string, opts tfe.WorkspaceLockOptions) (*tfe.Workspace, error) {
-	return &tfe.Workspace{
-		ID: "ws-123",
-	}, nil
-}
-
-func (f *FakeWorkspacesClient) Unlock(ctx context.Context, id string) (*tfe.Workspace, error) {
-	return &tfe.Workspace{
-		ID: "ws-123",
-	}, nil
 }
